fix: update custom hook and add to hooks/api/characters folder
