+++ src/core/hooks/index.ts
@@ -3,5 +3,3 @@ export * from './useLocalStorage'
 export * from './useRedirectFrom'
 export * from './useDebounce'
 export * from './useTimeout'
-
-export * from './useGetCharacters'
